Update parsers, SimpleRepo add, models __str__
DepartmentParser returns course code and title. Updated pytest
parameters to reflect change.

ScheduleParser has a check to see if row processed is an exam or
filler text. Uploaded BILD examples.

SimpleRepo add methods changed to return created object instead of
boolean. This is because we the IDs are needed when creating other
entries.

Implemented __str__ in models for debugging.